+++ src/mofaflex/_core/pyro/__init__.py
@@ -1,1 +1,3 @@
 from . import likelihoods, priors
+from .model import MofaFlexModel
+from .utils import PyroParameterDict
