+++ src/mofaflex/_core/pyro/utils.py
@@ -1,7 +1,11 @@
 # https://stackoverflow.com/a/61350480
 from abc import ABC
 
+import torch
 from pyro.nn import PyroModule
+
+PyroParameterDict = PyroModule[torch.nn.ParameterDict]
+PyroModuleDict = PyroModule[torch.nn.ModuleDict]
 
 
 class _PyroMeta(type(ABC), type(PyroModule)):
